# Feature/bloom optimizations