--- a/engine/renderer/private/passes/bloom_downsample_pass.cpp
+++ b/engine/renderer/private/passes/bloom_downsample_pass.cpp
@@ -19,7 +19,7 @@ BloomDownsamplePass::~BloomDownsamplePass()
     _context->VulkanContext()->Device().destroy(_pipelineLayout);
 }
 
-void BloomDownsamplePass::RecordCommands(vk::CommandBuffer commandBuffer, uint32_t currentFrame, const RenderSceneDescription& scene)
+void BloomDownsamplePass::RecordCommands(vk::CommandBuffer commandBuffer, MAYBE_UNUSED uint32_t currentFrame, const RenderSceneDescription& scene)
 {
     TracyVkZone(scene.tracyContext, commandBuffer, "Bloom Downsample Pass");