… engine/renderer/public/passes/bloom_downsample_pass.hpp
… engine/renderer/public/passes/bloom_upsample_pass.hpp
… engine/renderer/public/renderer.hpp
… engine/settings/public/settings.hpp
… settings.json
… shaders/bloom_downsample.frag
… shaders/bloom_downsampling.frag
… shaders/bloom_upsampling.frag
… shaders/settings.glsl