### Reviewer steps

- [ ] I can build the project locally
- [ ] I have tested and approved the features from this pull request
- [ ] I have checked and approved the test criteria
- [ ] I have reviewed the files in the pull request
- [ ] I have looked at and updated the related issues in Codecks

### Description

- Optimizes bloom rendering. Instead of gaussian blur, we down-sample and up-sample the bloom render target while using a specific sampling pattern.
- Removes the `brightness` render target as it was no longer needed.

Before:
- Full Frame | ~7.32ms
- Bloom Total (gaussian blur) |  ~2ms

![7-32FullFrameBefore](https://github.com/user-attachments/assets/3e4f1829-e08a-43c1-929c-4c0eb50baefd)

After:
- Full Frame | ~5.89ms
- Bloom Total (down-sampling and up-sampling) | ~0.3ms

![5-89msFullFrameAfter](https://github.com/user-attachments/assets/8e2d5701-1e13-43be-a916-2b407708dbf6)

Improvement:
- Full Frame Difference | ~1.43ms
- Bloom Total Difference | ~1.7ms
(Difference in these 2 is because the geometry pass took a bit longer during the 'after' test comparted to the 'before' test)

### Issues

[Codecks task](https://bubonic-brotherhood.codecks.io/milestones/run/58/card/1qa-optimizing-gaussian-blur)

### Test criteria

- [ ] The bloom renders correctly on the final image
- [ ] The bloom parameters still work in the imgui window
- [ ] Bloom renders faster than before